+++ app/javascript/react/components/Map/Markers/StreamMarkers.tsx
@@ -3,7 +3,7 @@ import React, { useEffect, useRef, useState } from "react";
 import { AdvancedMarker, useMap } from "@vis.gl/react-google-maps";
 
 import { mobileStreamPath, red } from "../../../assets/styles/colors";
-import { Session } from "./SessionType";
+import { Session } from "../../../types/sessionType";
 import { StreamMarker } from "./StreamMarker/StreamMarker";
 
 import type { Marker } from "@googlemaps/markerclusterer";
