+++ app/javascript/react/store/fixedSessionsSelectors.ts
@@ -1,6 +1,6 @@
 import { createSelector } from "reselect";
 
-import { Session } from "../components/Map/Markers/SessionType";
+import { Session } from "../types/sessionType";
 import { RootState } from "./";
 
 const selectFixedSessionsState = (state: RootState) => state.fixedSessions;
